+++ research/unity-humanoid-avatar-sdk/GakumasAvatarSdk/Assets/GakumasAvatarSdk/Runtime/VL/IKBodyEffector.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-namespace VL.IK
-{
-    public class IKBodyEffector : MonoBehaviour
-    {
-    }
-}